Repurpose the program to work with a graphical ui
Repurposed the program to work with a graphical version of the game in
the browser. Replaced the functionality of printing a board in the
console with one of generating a graphical board on the screen and
ensuring that each space therein is displays the marker of a player when
clicked.

Implemented a mode where human plays other human with each it being made
known who's turn it is to play each round via graphical marker atop the
game's board. Additionally Implemented the notion of a reset button
whereby the players can reset the game to the begin.